requirements.txt
# Email Automation System Dependencies

# Email handling
email-validator==2.1.0
python-dotenv==1.0.0

# LLM interaction
requests==2.31.0

# CLI interface
colorama==0.4.6
tabulate==0.9.0

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dateutil==2.8.2
